moved AuthServicesList into vsapi_types, added custom VisaRevocation
