+++ zpr/Cargo.toml
@@ -11,6 +11,7 @@ strum = { version = "0.26", features = ["derive"] }
 zerocopy = { version = "0.8", features = ["derive"] }
 thiserror = "1.0.65"
 vsapi = { git = "https://github.com/org-zpr/zpr-vsapi-rs.git", tag = "v0.4.1" }
+url = "2.5.4"
 
 [build-dependencies]
 capnpc = "0.23"
